Properly handle telnet transport for Dell switches.
Also included:
 - a refactor of the way command processors interact with the terminal.
 - split of Dell switch tests in multiple modules
 - Dell switch tests on telnet
 - a new way of handling "press any key to continue"
