+++ tests/dell/test_configure_interface.py
@@ -1,440 +1,39 @@
-import pprint
+# Copyright 2015 Internap.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import unittest
 
 from flexmock import flexmock_teardown
-from hamcrest import assert_that, is_, has_item, is_not
+from hamcrest import assert_that, is_not, has_item
 
-from tests.util.global_reactor import dell_privileged_password
-from tests.util.global_reactor import dell_switch_ip, dell_switch_ssh_port
-from tests.util.protocol_util import SshTester, with_protocol
+from tests.dell import enable, configuring_interface, \
+    assert_interface_configuration, assert_running_config_contains_in_order, \
+    get_running_config, configure, configuring_vlan, unconfigure_vlan, \
+    configuring_a_vlan_on_interface, create_bond, remove_bond, \
+    ssh_protocol_factory, telnet_protocol_factory
+from tests.util.protocol_util import with_protocol
 
 
-class DellSwitchTest(unittest.TestCase):
+class DellConfigureInterfaceTest(unittest.TestCase):
     __test__ = False
+    protocol_factory = None
 
     def setUp(self):
-        self.protocol = SshTester("ssh", dell_switch_ip, dell_switch_ssh_port, 'root', 'root')
+        self.protocol = self.protocol_factory()
 
     def tearDown(self):
         flexmock_teardown()
-
-
-class DellUnprivilegedProtocolTest(DellSwitchTest):
-    __test__ = True
-
-    @with_protocol
-    def test_entering_enable_mode_requires_a_password(self, t):
-        t.write("enable")
-        t.read("Password:")
-        t.write_stars(dell_privileged_password)
-        t.read("\r\n")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_wrong_password(self, t):
-        t.write("enable")
-        t.read("Password:")
-        t.write_stars("hello_world")
-        t.readln("Incorrect Password!")
-        t.read("my_switch>")
-
-    @with_protocol
-    def test_no_password_works_for_legacy_reasons(self, t):
-        t.write("enable")
-        t.read("Password:")
-        t.write_stars("")
-        t.read("\r\n")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_exit_disconnects(self, t):
-        t.write("exit")
-        t.read_eof()
-
-    @with_protocol
-    def test_quit_disconnects(self, t):
-        t.write("quit")
-        t.read_eof()
-
-
-class DellEnabledProtocolTest(DellSwitchTest):
-    __test__ = True
-
-    @with_protocol
-    def test_exit_returns_to_unprivileged_mode(self, t):
-        enable(t)
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch>")
-
-    @with_protocol
-    def test_quit_disconnects(self, t):
-        enable(t)
-        t.write("quit")
-        t.read_eof()
-
-    @with_protocol
-    def test_write_memory(self, t):
-        enable(t)
-
-        t.write("copy running-config startup-config")
-
-        t.readln("")
-        t.readln("This operation may take a few minutes.")
-        t.readln("Management interfaces will not be available during this time.")
-        t.readln("")
-        t.read("Are you sure you want to save? (y/n) ")
-        t.write_raw("y")
-        t.read("y")
-        t.readln("")
-        t.readln("")
-        t.readln("Configuration Saved!")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_write_memory_abort(self, t):
-        enable(t)
-
-        t.write("copy running-config startup-config")
-
-        t.readln("")
-        t.readln("This operation may take a few minutes.")
-        t.readln("Management interfaces will not be available during this time.")
-        t.readln("")
-        t.read("Are you sure you want to save? (y/n) ")
-        t.write_raw("n")
-        t.read("n")
-        t.readln("")
-        t.readln("")
-        t.readln("Configuration Not Saved!")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_write_memory_any_other_key_aborts(self, t):
-        enable(t)
-
-        t.write("copy running-config startup-config")
-
-        t.readln("")
-        t.readln("This operation may take a few minutes.")
-        t.readln("Management interfaces will not be available during this time.")
-        t.readln("")
-        t.read("Are you sure you want to save? (y/n) ")
-        t.write_raw("p")
-        t.read("p")
-        t.readln("")
-        t.readln("")
-        t.readln("Configuration Not Saved!")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_invalid_command(self, t):
-        enable(t)
-
-        t.write("shizzle")
-        t.readln("          ^")
-        t.readln("% Invalid input detected at '^' marker.")
-        t.readln("")
-
-    @with_protocol
-    def test_entering_configure_mode(self, t):
-        enable(t)
-
-        t.write("configure")
-        t.readln("")
-        t.read("my_switch(config)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_show_running_config_on_empty_ethernet_port(self, t):
-        enable(t)
-
-        t.write("show running-config interface ethernet 1/g1")
-        t.readln("")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_show_running_config_on_ethernet_port_that_does_not_exists(self, t):
-        enable(t)
-
-        t.write("show running-config interface ethernet 4/g8")
-        t.readln("")
-        t.read("ERROR: Invalid input!")
-        t.readln("")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_show_running_config_displays_header(self, t):
-        enable(t)
-        assert_running_config_contains_in_order(t, [
-            '!Current Configuration:',
-            '!System Description "PowerConnect 6224P, 3.3.7.3, VxWorks 6.5"',
-            '!System Software Version 3.3.7.3',
-            '!Cut-through mode is configured as disabled',
-            '!',
-            'configure',
-        ])
-
-    @with_protocol
-    def test_show_vlan(self, t):
-        enable(t)
-
-        configuring_vlan(t, 10)
-        configuring_vlan(t, 11)
-        configuring_vlan(t, 12)
-        configuring_vlan(t, 13)
-        configuring_vlan(t, 14)
-        configuring_vlan(t, 15)
-        configuring_vlan(t, 16)
-        configuring_vlan(t, 17)
-        configuring_interface_vlan(t, 17, do="name this-name-is-too-long-buddy-budd")
-        configuring_vlan(t, 18)
-        configuring_vlan(t, 19)
-        configuring_vlan(t, 20)
-        configuring_vlan(t, 21)
-        configuring_vlan(t, 22)
-        configuring_vlan(t, 23)
-        configuring_vlan(t, 24)
-        configuring_vlan(t, 25)
-        configuring_vlan(t, 26)
-        configuring_vlan(t, 27)
-        configuring_vlan(t, 28)
-        configuring_vlan(t, 29)
-        configuring_vlan(t, 300)
-        configuring_vlan(t, 4000)
-        configuring_interface_vlan(t, 300, do="name shizzle")
-
-        t.write("show vlan")
-        t.readln("")
-        t.readln("VLAN       Name                         Ports          Type      Authorization")
-        t.readln("-----  ---------------                  -------------  -----     -------------")
-        t.readln("1      Default                                         Default   Required     ")
-        t.readln("10                                                     Static    Required     ")
-        t.readln("11                                                     Static    Required     ")
-        t.readln("12                                                     Static    Required     ")
-        t.readln("13                                                     Static    Required     ")
-        t.readln("14                                                     Static    Required     ")
-        t.readln("15                                                     Static    Required     ")
-        t.readln("16                                                     Static    Required     ")
-        t.readln("17     this-name-is-too-long-buddy-budd                Static    Required     ")
-        t.readln("18                                                     Static    Required     ")
-        t.readln("19                                                     Static    Required     ")
-        t.readln("20                                                     Static    Required     ")
-        t.readln("21                                                     Static    Required     ")
-        t.readln("22                                                     Static    Required     ")
-        t.readln("23                                                     Static    Required     ")
-        t.readln("24                                                     Static    Required     ")
-        t.readln("25                                                     Static    Required     ")
-        t.readln("26                                                     Static    Required     ")
-        t.readln("")
-        t.read("--More-- or (q)uit")
-        t.write_raw("m")
-        t.read("m")
-        t.readln("\r                     ")
-        t.readln("")
-        t.readln("")
-        t.readln("VLAN       Name                         Ports          Type      Authorization")
-        t.readln("-----  ---------------                  -------------  -----     -------------")
-        t.readln("27                                                     Static    Required     ")
-        t.readln("28                                                     Static    Required     ")
-        t.readln("29                                                     Static    Required     ")
-        t.readln("300    shizzle                                         Static    Required     ")
-        t.readln("4000                                                   Static    Required     ")
-        t.readln("")
-        t.read("my_switch#")
-
-        unconfigure_vlan(t, 10)
-        unconfigure_vlan(t, 11)
-        unconfigure_vlan(t, 12)
-        unconfigure_vlan(t, 13)
-        unconfigure_vlan(t, 14)
-        unconfigure_vlan(t, 15)
-        unconfigure_vlan(t, 16)
-        unconfigure_vlan(t, 17)
-        unconfigure_vlan(t, 18)
-        unconfigure_vlan(t, 19)
-        unconfigure_vlan(t, 20)
-        unconfigure_vlan(t, 21)
-        unconfigure_vlan(t, 22)
-        unconfigure_vlan(t, 23)
-        unconfigure_vlan(t, 24)
-        unconfigure_vlan(t, 25)
-        unconfigure_vlan(t, 26)
-        unconfigure_vlan(t, 27)
-        unconfigure_vlan(t, 28)
-        unconfigure_vlan(t, 29)
-        unconfigure_vlan(t, 300)
-        unconfigure_vlan(t, 4000)
-
-
-class DellConfigureProtocolTest(DellSwitchTest):
-    __test__ = True
-
-    @with_protocol
-    def test_entering_configure_interface_mode(self, t):
-        enable(t)
-        configure(t)
-
-        t.write("interface ethernet 1/g1")
-        t.readln("")
-        t.read("my_switch(config-if-1/g1)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch(config)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_entering_vlan_database_mode(self, t):
-        enable(t)
-        configure(t)
-
-        t.write("vlan database")
-        t.readln("")
-        t.read("my_switch(config-vlan)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch(config)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_editing_vlan(self, t):
-        enable(t)
-        configure(t)
-
-        t.write("interface vlan 1260")
-        t.readln("VLAN ID not found.")
-        t.readln("")
-
-        t.read("my_switch(config)#")
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-        configuring_vlan(t, 1260)
-
-        configure(t)
-        t.write("interface vlan 1260")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1260)#")
-
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch(config)#")
-
-        unconfigure_vlan(t, 1260)
-
-
-class DellConfigureVlanProtocolTest(DellSwitchTest):
-    __test__ = True
-
-    @with_protocol
-    def test_configuring_a_vlan(self, t):
-        enable(t)
-
-        configuring_vlan(t, 1234)
-
-        assert_running_config_contains_in_order(t, [
-            "vlan database",
-            "vlan 1,1234",
-            "exit"
-        ])
-
-        unconfigure_vlan(t, 1234)
-
-        assert_running_config_contains_in_order(t, [
-            "vlan database",
-            "vlan 1",
-            "exit"
-        ])
-
-    @with_protocol
-    def test_unconfiguring_a_vlan_failing(self, t):
-        enable(t)
-
-        t.write("configure")
-        t.readln("")
-        t.read("my_switch(config)#")
-
-        t.write("vlan database")
-        t.readln("")
-        t.read("my_switch(config-vlan)#")
-
-        t.write("no vlan 3899")
-        t.readln("Warning: The use of large numbers of VLANs or interfaces may cause significant")
-        t.readln("delays in applying the configuration.")
-        t.readln("")
-        t.readln("")
-        t.readln("These VLANs do not exist:  3899.")
-        t.readln("")
-
-        t.read("my_switch(config-vlan)#")
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch(config)#")
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-    @with_protocol
-    def test_configuring_a_vlan_name(self, t):
-        enable(t)
-
-        configuring_vlan(t, 1234)
-
-        t.write("configure")
-        t.readln("")
-        t.read("my_switch(config)#")
-        t.write("interface vlan 1234")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1234)#")
-        t.write("name")
-        t.readln("")
-        t.readln("Command not found / Incomplete command. Use ? to list commands.")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1234)#")
-        t.write("name one two")
-        t.readln("                                     ^")
-        t.readln("% Invalid input detected at '^' marker.")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1234)#")
-        t.write("name this-name-is-too-long-buddy-buddy")
-        t.readln("Name must be 32 characters or less.")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1234)#")
-        t.write("name this-name-is-too-long-buddy-budd")
-        t.readln("")
-        t.read("my_switch(config-if-vlan1234)#")
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch(config)#")
-        t.write("exit")
-        t.readln("")
-        t.read("my_switch#")
-
-        assert_interface_configuration(t, "vlan 1234", [
-            "interface vlan 1234",
-            "name \"this-name-is-too-long-buddy-budd\"",
-            "exit",
-        ])
-
-        unconfigure_vlan(t, 1234)
-
-
-class DellConfigureInterfaceProtocolTest(DellSwitchTest):
-    __test__ = True
 
     @with_protocol
     def test_show_run_vs_show_run_interface_same_output(self, t):
@@ -1079,7 +678,6 @@ class DellConfigureInterfaceProtocolTest(DellSwitchTest):
         t.readln("ch8  Link Aggregate                  Down")
         t.read("--More-- or (q)uit")
         t.write_raw("m")
-        t.read("m")
         t.readln("")
         t.readln("ch9  Link Aggregate                  Down")
         t.readln("ch10 Link Aggregate                  Down")
@@ -1100,190 +698,11 @@ class DellConfigureInterfaceProtocolTest(DellSwitchTest):
         remove_bond(t, 10)
 
 
-def enable(t):
-    t.write("enable")
-    t.read("Password:")
-    t.write_stars(dell_privileged_password)
-    t.readln("")
-    t.read("my_switch#")
+class DellConfigureInterfaceSshTest(DellConfigureInterfaceTest):
+    __test__ = True
+    protocol_factory = ssh_protocol_factory
 
 
-def configure(t):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-
-
-def configuring_interface(t, interface, do):
-    interface_short_name = interface.split(' ')[1]
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("interface %s" % interface)
-    t.readln("")
-    t.read("my_switch(config-if-%s)#" % interface_short_name)
-
-    t.write(do)
-
-    t.readln("")
-    t.read("my_switch(config-if-%s)#" % interface_short_name)
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def configuring_a_vlan_on_interface(t, interface, do):
-    interface_short_name = interface.split(' ')[1]
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("interface %s" % interface)
-    t.readln("")
-    t.read("my_switch(config-if-%s)#" % interface_short_name)
-
-    t.write(do)
-
-    t.readln("Warning: The use of large numbers of VLANs or interfaces may cause significant")
-    t.readln("delays in applying the configuration.")
-    t.readln("")
-    t.readln("")
-    t.read("my_switch(config-if-%s)#" % interface_short_name)
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def configuring_interface_vlan(t, vlan, do):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("interface vlan {}".format(vlan))
-    t.readln("")
-    t.read("my_switch(config-if-vlan{})#".format(vlan))
-
-    t.write(do)
-
-    t.readln("")
-    t.read("my_switch(config-if-vlan{})#".format(vlan))
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def configuring_vlan(t, vlan_id):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-
-    t.write("vlan database")
-    t.readln("")
-    t.read("my_switch(config-vlan)#")
-
-    t.write("vlan %s" % vlan_id)
-    t.readln("Warning: The use of large numbers of VLANs or interfaces may cause significant")
-    t.readln("delays in applying the configuration.")
-    t.readln("")
-
-    t.readln("")
-    t.read("my_switch(config-vlan)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def unconfigure_vlan(t, vlan_id):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-
-    t.write("vlan database")
-    t.readln("")
-    t.read("my_switch(config-vlan)#")
-
-    t.write("no vlan %s" % vlan_id)
-    t.readln("Warning: The use of large numbers of VLANs or interfaces may cause significant")
-    t.readln("delays in applying the configuration.")
-    t.readln("")
-    t.readln("If any of the VLANs being deleted are for access ports, the ports will be")
-    t.readln("unusable until it is assigned a VLAN that exists.")
-    t.readln("")
-
-    t.read("my_switch(config-vlan)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def create_bond(t, bond_id):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("interface port-channel {}".format(bond_id))
-    t.readln("")
-    t.read("my_switch(config-if-ch{})#".format(bond_id))
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def remove_bond(t, bond_id):
-    t.write("configure")
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("backdoor remove port-channel {}".format(bond_id))
-    t.readln("")
-    t.read("my_switch(config)#")
-    t.write("exit")
-    t.readln("")
-    t.read("my_switch#")
-
-
-def assert_interface_configuration(t, interface, config):
-    t.write("show running-config interface %s " % interface)
-    for line in config:
-        t.readln(line)
-    t.readln("")
-    t.read("my_switch#")
-
-
-def assert_running_config_contains_in_order(t, lines):
-    config = get_running_config(t)
-
-    assert_lines_order(config, lines)
-
-
-def get_running_config(t):
-    t.write("show running-config")
-    config = t.read_lines_until('my_switch#')
-    return config
-
-
-def assert_lines_order(config, lines):
-    begin = config.index(lines[0])
-
-    for (i, line) in enumerate(lines):
-        expected_content = line
-        expected_line_number = i + begin
-        actual_content = config[expected_line_number]
-
-        assert_that(actual_content, is_(expected_content),
-                    "Item <%s> was expected to be found at line %s but found %s instead.\nWas looking for %s in %s" % (
-                    line, expected_line_number, actual_content, pprint.pformat(lines), pprint.pformat(config)))
+class DellConfigureInterfaceTelnetTest(DellConfigureInterfaceTest):
+    __test__ = True
+    protocol_factory = telnet_protocol_factory
